CORDA-3264 Split up sweepstake app into states/ flows/ cordapp modules to avoid cyclic dependencies among other side effects:
