+++ examples/cordapp-sweepstake/cordapp/build.gradle
@@ -12,27 +12,47 @@ cordapp {
     }
 }
 
+
+sourceSets {
+    integrationTest {
+        kotlin {
+            compileClasspath += main.output + test.output
+            runtimeClasspath += main.output + test.output
+            srcDir file('src/integrationTest/kotlin')
+        }
+    }
+}
+
+configurations {
+    integrationTestCompile.extendsFrom testCompile
+    integrationTestRuntime.extendsFrom testRuntime
+}
+
+
 dependencies {
     cordapp project(":workflows")
     cordapp project(":contracts")
-    cordapp project(":examples:cordapp-sweepstake:states-and-flows")
+    cordapp project(":examples:cordapp-sweepstake:states")
+    cordapp project(":examples:cordapp-sweepstake:flows")
 
     cordaRuntime "$corda_release_group:corda:$corda_release_version"
     cordaCompile "$corda_release_group:corda-core:$corda_release_version"
 
-    cordapp "$tokens_release_group:tokens-contracts:$tokens_release_version"
-    cordapp "$tokens_release_group:tokens-workflows:$tokens_release_version"
-    cordapp "$tokens_release_group:tokens-money:$tokens_release_version"
+    cordaCompile "$tokens_release_group:tokens-contracts:$tokens_release_version"
+    cordaCompile "$tokens_release_group:tokens-workflows:$tokens_release_version"
+    cordaCompile "$tokens_release_group:tokens-money:$tokens_release_version"
 
     cordaCompile "$corda_release_group:corda-rpc:$corda_release_version"
     cordaCompile "$corda_release_group:corda-jackson:$corda_release_version"
-
     compile("org.springframework.boot:spring-boot-starter-websocket:$spring_boot_version") {
         exclude group: "org.springframework.boot", module: "spring-boot-starter-logging"
     }
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    compile "com.beust:klaxon:5.0.1"
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
 
+    testCompile project(":examples:cordapp-sweepstake:states")
+    testCompile project(":examples:cordapp-sweepstake:flows")
     testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
     testCompile "junit:junit:$junit_version"
     testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
@@ -49,7 +69,8 @@ task deployNodesX(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
         }
         cordapp project(":workflows")
         cordapp project(":contracts")
-        cordapp project(":examples:cordapp-sweepstake:states-and-flows")
+        cordapp project(":examples:cordapp-sweepstake:states")
+        cordapp project(":examples:cordapp-sweepstake:flows")
         cordapp "$tokens_release_group:tokens-contracts:$tokens_release_version"
         cordapp "$tokens_release_group:tokens-workflows:$tokens_release_version"
         cordapp "$tokens_release_group:tokens-money:$tokens_release_version"
@@ -86,4 +107,15 @@ task deployNodesX(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
         }
         rpcUsers = [[user: "user1", "password": "test", "permissions": ["ALL"]]]
     }
+}
+
+task integrationTest(type: Test, dependsOn: []) {
+    testClassesDirs = sourceSets.integrationTest.output.classesDirs
+    classpath = sourceSets.integrationTest.runtimeClasspath
+}
+
+task runTournamentServer(type: JavaExec, dependsOn: assemble) {
+    classpath = sourceSets.main.runtimeClasspath
+    main = 'com.r3.corda.lib.accounts.examples.sweepstake.clients.WorldCupApplication'
+    args '--server.port=8080', '--config.rpc.host=localhost', '--config.rpc.port=10005', '--config.rpc.username=user1', '--config.rpc.password=test'
 }
