+++ examples/cordapp-sweepstake/cordapp/src/main/resources/application.properties
