+++ examples/cordapp-sweepstake/cordapp/src/main/resources/participants.txt
