+++ examples/cordapp-sweepstake/cordapp/src/main/resources/static/index.html
