+++ examples/cordapp-sweepstake/cordapp/src/main/resources/static/sweepstake.css
