+++ examples/cordapp-sweepstake/cordapp/src/main/resources/static/sweepstake.js
