+++ examples/cordapp-sweepstake/cordapp/src/main/resources/worldcupteams.txt
