+++ examples/cordapp-sweepstake/flows/build.gradle
@@ -1,0 +1,45 @@
+apply plugin: 'kotlin'
+apply plugin: 'net.corda.plugins.cordapp'
+apply plugin: 'net.corda.plugins.cordformation'
+apply plugin: 'net.corda.plugins.quasar-utils'
+
+
+cordapp {
+    info {
+        name "World Cup Sweepstake - Workflows"
+        vendor "Corda Open Source"
+        targetPlatformVersion 4
+        minimumPlatformVersion 4
+    }
+    signing {
+        enabled false
+    }
+    sealing {
+        enabled false
+    }
+}
+
+
+dependencies {
+    // Accounts.
+    cordapp project(":workflows")
+//    cordapp project(":contracts")
+    cordapp project(":examples:cordapp-sweepstake:states")
+
+    // Corda.
+    cordaCompile "$corda_release_group:corda-core:$corda_release_version"
+
+    // Tokens.
+    cordapp "$tokens_release_group:tokens-contracts:$tokens_release_version"
+    cordapp "$tokens_release_group:tokens-workflows:$tokens_release_version"
+    cordapp "$tokens_release_group:tokens-money:$tokens_release_version"
+
+    // Kotlin.
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+
+    // Testing.
+    testCompile project(":examples:cordapp-sweepstake:states")
+    testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
+    testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
+    testCompile "junit:junit:$junit_version"
+}
