+++ examples/cordapp-sweepstake/flows/src/main/resources/participants.txt
@@ -1,0 +1,32 @@
+Oliver
+Jack
+Harry
+Jacob
+Charlie
+Thomas
+George
+Oscar
+James
+William
+Noah
+Alfie
+Joshua
+Muhammad
+Henry
+Leo
+Amelia
+Olivia
+Isla
+Emily
+Poppy
+Ava
+Isabella
+Jessica
+Lily
+Sophie
+Grace
+Sophia
+Mia
+Evie
+Ruby
+Ella
