+++ examples/cordapp-sweepstake/flows/src/main/resources/worldcupteams.txt
@@ -1,0 +1,32 @@
+Russia
+Saudi Arabia
+Egypt
+Uruguay
+Portugal
+Spain
+Morocco
+Iran
+France
+Australia
+Peru
+Denmark
+Argentina
+Iceland
+Croatia
+Nigeria
+Brazil
+Switzerland
+Costa Rica
+Serbia
+Germany
+Mexico
+Sweden
+South Korea
+Belgium
+Panama
+Tunisia
+England
+Poland
+Senegal
+Colombia
+Japan
