+++ examples/cordapp-sweepstake/states/build.gradle
@@ -1,0 +1,25 @@
+apply plugin: 'kotlin'
+apply plugin: 'net.corda.plugins.cordapp'
+apply plugin: 'net.corda.plugins.cordformation'
+apply plugin: 'net.corda.plugins.quasar-utils'
+
+
+cordapp {
+    info {
+        name "World Cup Sweepstake - Contracts"
+        vendor "Corda Open Source"
+        targetPlatformVersion 4
+        minimumPlatformVersion 4
+    }
+    signing {
+        enabled false
+    }
+}
+
+dependencies {
+    // Corda.
+    cordaCompile "$corda_release_group:corda-core:$corda_release_version"
+
+    // Kotlin.
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+}
